Iteration_3/python_proj/requirements.txt
fastapi
llama_cpp-python
sse_starlette
httpx
uvicorn
